commit: Store objtype in size index
If the objtype is not stored in the size index, the index will
be ambiguous if the size of anything other than regular files
is stored.

This adds the objtype in a single byte following the existing data,
ensuring that code designed to read the old format will continue
to work with this addition.