commit: Implement ostree_repo_get_commit_sizes to read size index
There is currently no code to read the size index. Add a function
ostree_repo_get_commit_sizes to read the total and new size of the
objects in a given commit in terms of both archived size (i.e.
download size) and unpacked size (i.e. on disk size).