lib/remote: Add a method to return the URL
This method allows to return the URL directly from the OstreeRemote
without having to go through the OstreeRepo. This is needed e.g. for
getting the URL from temporary remotes (USB) since they are not stored
in an OstreeRepo object.

The test-repo-finder-config is also updated by this patch to check if
the URL is correct.